move `Quad`, `Rect` and `BBox` to `gcore::math`
