+++ editor/src/messages/portfolio/document/node_graph/node_graph_message_handler.rs
@@ -21,6 +21,7 @@ use crate::messages::tool::utility_types::{HintData, HintGroup, HintInfo};
 use glam::{DAffine2, DVec2, IVec2};
 use graph_craft::document::{DocumentNodeImplementation, NodeId, NodeInput};
 use graph_craft::proto::GraphErrors;
+use graphene_std::math::math_ext::QuadExt;
 use graphene_std::*;
 use renderer::Quad;
 use std::cmp::Ordering;
