+++ node-graph/gcore/src/math/bbox.rs
