+++ node-graph/gcore/src/vector/brush_stroke.rs
@@ -1,6 +1,6 @@
 use crate::Color;
+use crate::math::bbox::AxisAlignedBbox;
 use crate::raster::BlendMode;
-use crate::raster::bbox::AxisAlignedBbox;
 use dyn_any::DynAny;
 use glam::DVec2;
 use std::hash::{Hash, Hasher};
