+++ node-graph/gstd/src/brush.rs
@@ -3,8 +3,8 @@ use glam::{DAffine2, DVec2};
 use graph_craft::generic::FnNode;
 use graph_craft::proto::FutureWrapperNode;
 use graphene_core::instances::Instance;
+use graphene_core::math::bbox::{AxisAlignedBbox, Bbox};
 use graphene_core::raster::adjustments::blend_colors;
-use graphene_core::raster::bbox::{AxisAlignedBbox, Bbox};
 use graphene_core::raster::brush_cache::BrushCache;
 use graphene_core::raster::image::Image;
 use graphene_core::raster::{Alpha, BitmapMut, BlendMode, Color, Pixel, Sample};
